turning the form and render into components and putting on one page
